Implement changes
